pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zfsnap"
version = "1.0.2"
authors = [
  { name="pschlo", email="author@example.com" },
]
description = "Policy-based ZFS snapshot pruning"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "python-dateutil"
]


[tool.hatch.build.targets.sdist]
# choose here what to include in sdist
# the wheel will only include src/zfsnap anyway
# see https://discuss.python.org/t/should-sdists-include-docs-and-tests/14578
# include = [ ]
